+++ src/diffusers/pipelines/stable_video_diffusion/pipeline_stable_video_diffusion.py
